refactor: :recycle: Déplacer les fichiers de pages vers le dossier components et ajuster les chemins d'importation
