+++ src/components/pages/LegalNotices/LegalNotices.tsx
